requirements.txt
# DiscoBall Python Tools Dependencies

# Web3 and Ethereum interaction
web3>=6.0.0
eth-account>=0.9.0

# IPFS client
ipfshttpclient>=0.8.0

# Web scraping and parsing
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# DNS resolution
dnspython>=2.3.0

# Type hints support for older Python versions
typing-extensions>=4.0.0

# Optional: Better JSON handling
orjson>=3.8.0 